Add env validation tests and resource planning docs
